% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy_Calc_2Class.R
\name{calculate.accuracy}
\alias{calculate.accuracy}
\title{Accuracy calculation for a 2 class model}
\usage{
calculate.accuracy(predictions, ref.labels)
}
\arguments{
\item{predictions}{Spectra data that has been interpolated already}

\item{ref.labels}{Reference labels for the spectra}
}
\description{
Accuracy calculation for a 2 class model
}
