% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy_Calc_Greater_Than_2Class.R
\name{calculate.w.accuracy}
\alias{calculate.w.accuracy}
\title{Accuracy calculation for a > 2 class model}
\usage{
calculate.w.accuracy(predictions, ref.labels, weights)
}
\arguments{
\item{predictions}{Spectra data that has been interpolated already}

\item{ref.labels}{Reference labels for the spectra}

\item{weights}{Weights for the classes}
}
\description{
Accuracy calculation for a > 2 class model
}
