% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence_Ellipse.R
\name{conf_ellipse_outliers}
\alias{conf_ellipse_outliers}
\title{Confidence ellipse and outliers from PC1/PC2}
\usage{
conf_ellipse_outliers(spectra, confidence)
}
\arguments{
\item{spectra}{Spectra data that has been interpolated already}

\item{confidence}{Chosen confidence level, recommend 95%}
}
\value{
List of 3 elements: Matrix with PC1/PC2 of your data,
Confidence ellipse matrix, the indices of any outliers
}
\description{
For computing a confidence ellipse for spectra pc1/pc2
can use with either raw or processed but recommend processed
since most of the variance occurs in the intensities
}
