% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMSC_Wrapper.R
\name{EMSC.PPRAMAN}
\alias{EMSC.PPRAMAN}
\title{EMSC wrapper - NOT TESTED USE WITH CAUTION}
\usage{
EMSC.PPRAMAN(spectra, background_spectra, reference_spectra)
}
\arguments{
\item{spectra}{Matrix of spectral data to have EMSC applied.}

\item{background_spectra}{Matrix of spectra to be used as the background component for EMSC. This is then averaged and smoothed.}

\item{reference_spectra}{Matrix of spectra to be used as the reference component for EMSC. This is then averaged and smoothed.}
}
\description{
EMSC wrapper - NOT TESTED USE WITH CAUTION
}
