% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_Importance.R
\name{feature_selection}
\alias{feature_selection}
\title{Feature Selection for RF}
\usage{
feature_selection(rf_model)
}
\arguments{
\item{rf_model}{Input a random forest model}
}
\value{
rf_imp
}
\description{
Short function to enable rapid extraction of random forest importnace for features
}
