% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flip_Data.R
\name{flip_data}
\alias{flip_data}
\title{Flip data}
\usage{
flip_data(files)
}
\arguments{
\item{files}{Raw input files}
}
\value{
The input files flipped from low to high
}
\description{
Flips the data so it runs from low to high
}
