% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMSC_Spectra_Formatting.R
\name{format_spectra}
\alias{format_spectra}
\title{EMSC Spectral Formatting}
\usage{
format_spectra(raw_spec, n_pat, n_reps)
}
\arguments{
\item{raw_spec}{Spectra data that has been interpolated already}

\item{n_pat}{Number of patients - user input}

\item{n_reps}{Number of repeats for each patient}
}
\value{
Formatted spectra for EMSC
}
\description{
Formats spectra from hyperspec to something consistent with EMSC pkg
}
