% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPC_Read_in_interp.R
\name{hpc_read_in_interp}
\alias{hpc_read_in_interp}
\title{HPC READ IN AND INTERPOLATE SPECTRA}
\usage{
hpc_read_in_interp(spectra)
}
\description{
Reads in spectra, shifts to 1004 and interpolates on the x axis
Note currently phenyl peak to 2nd order poly rather than 4th due to error
}
