% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolate_Only.R
\name{interpolate_spectra}
\alias{interpolate_spectra}
\title{Interpolate Spectra}
\usage{
interpolate_spectra()
}
\value{
Interpolated spectra using approx function method = "constant"
}
\description{
Interpolates raw spectral data using Approx function, no smoothing
}
