% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Is_nan.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Replace NaN with 0
Function for replacing NaN's with zeros in the entire df}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{Data frame / structure}
}
\value{
Data frame with all NaN's replaced with 0's
}
\description{
Replace NaN with 0
Function for replacing NaN's with zeros in the entire df
}
