% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_MinMax.R
\name{norm_minmax}
\alias{norm_minmax}
\title{Normalise using min max}
\usage{
norm_minmax(spectra)
}
\arguments{
\item{spectra}{Spectra in matrix format for normalising}
}
\value{
Normalised spectra
}
\description{
Normalising function using min/max method
}
