% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Phenylalanine.R
\name{norm_p}
\alias{norm_p}
\title{Normalise at Phenylalanine peak}
\usage{
norm_p(spectra)
}
\arguments{
\item{spectra}{Spectra in matrix format for normalising}
}
\value{
Normalised spectra
}
\description{
Normalising function using min/max method at phenylalanine peak
}
