% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_RNV.R
\name{norm_rnv}
\alias{norm_rnv}
\title{Normalising Spectra using RNV}
\usage{
norm_rnv(spectra)
}
\arguments{
\item{spectra}{Spectral like data}
}
\value{
Normalised spectra using RNV method
}
\description{
Normalises spectra using Robust Normal Variate method - RNV was introduced to solve
a closure problem - in statistics this is
}
