% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_SNV.R
\name{norm_snv}
\alias{norm_snv}
\title{Normalising Spectra using SNV}
\usage{
norm_snv(spectra)
}
\arguments{
\item{spectra}{Spectral like data}
}
\description{
Normalises spectra using Standard Normal Variate method
}
