% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Vec.R
\name{norm_vec}
\alias{norm_vec}
\title{Normalising Spectra using Vector Normalising}
\usage{
norm_vec(spectra)
}
\arguments{
\item{spectra}{Spectral like data}
}
\description{
Normalising Spectra using Vector Normalising
}
