% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options_Func.R
\name{opt_process_hpc}
\alias{opt_process_hpc}
\title{Options for pre-processing spectra}
\usage{
opt_process_hpc(spectra, norm_meth, RCF_rad)
}
\arguments{
\item{spectra}{Spectra data that has been interpolated already}

\item{norm_meth}{Choice of normalisation}

\item{RCF_rad}{RCF radius selection}
}
\value{
Preprocessed spectra
}
\description{
Options for pre-processing spectra
}
