% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_func.R
\name{opt_process_hpc_dev}
\alias{opt_process_hpc_dev}
\title{Options for pre-processing spectra - developing version}
\usage{
opt_process_hpc_dev(spectra, norm_meth, rm_bl, baseline_fit, RCF_rad,
  poly_order)
}
\arguments{
\item{spectra}{Spectra data that has been interpolated already}

\item{norm_meth}{Choice of normalisation}

\item{rm_bl}{Choice of background removal}

\item{baseline_fit}{Baseline fit for RCF - currently off}

\item{RCF_rad}{RCF radius selection}

\item{poly_order}{Choice of polynomial order for SavGol etc}
}
\value{
Preprocessed spectra
}
\description{
Options for pre-processing spectra - developing version
}
