% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opt_Process_Spec_Testing_Dev.R
\name{opt_process_hpc_eff}
\alias{opt_process_hpc_eff}
\title{Options V1 Norm & RF Radius efficiency testing
First run of options for the HPC -- testing phase}
\usage{
opt_process_hpc_eff(spectra, norm_meth, RCF_rad)
}
\arguments{
\item{spectra}{Spectra data that has been interpolated already}

\item{norm_meth}{Choice of normalisation}

\item{RCF_rad}{RCF radius selection}
}
\value{
Preprocessed spectra
}
\description{
Options V1 Norm & RF Radius efficiency testing
First run of options for the HPC -- testing phase
}
