% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opt_Process_HPC_Smooth.R
\name{opt_process_hpc_smooth}
\alias{opt_process_hpc_smooth}
\title{Options for pre-processing spectra - developing version}
\usage{
opt_process_hpc_smooth(spectra, norm_meth, rm_bl, RCF_rad, poly_bl_order,
  filter_length, deriv_order = 0, poly_smooth_order, smoothing_opt = 0,
  binning_opt = 0, bin_size = 1)
}
\arguments{
\item{spectra}{Spectra data that has been interpolated already}

\item{norm_meth}{Choice of normalisation}

\item{rm_bl}{Choice of background removal}

\item{RCF_rad}{RCF radius selection}

\item{poly_bl_order}{Choice of polynomial order for baseline removal}

\item{filter_length}{filter length (window) for savitsky golay smoothing}

\item{deriv_order}{derivative order choice for baseline correction}

\item{poly_smooth_order}{polynomial order for smoothing}

\item{smoothing_opt}{Binary 1 or 0 if you want smoothing or not! This also automatically removes the dud wavenumbers from the fl + 1/2 at either end}

\item{binning_opt}{Binary 1 or 0 for binning option}

\item{bin_size}{A numeric value for the desired # of wavenumbers in each bin}
}
\value{
Preprocessed spectra
}
\description{
Options for pre-processing spectra - developing version
}
