% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patient_Result_Coordinator.R
\name{patient_results}
\alias{patient_results}
\title{Patient diagnosis}
\usage{
patient_results(patient_ids, clinical_result, raman_result)
}
\arguments{
\item{patient_ids}{Vector containing patient IDs}

\item{clinical_result}{Vector containing binary clinical result}

\item{raman_result}{Vector containing binary raman result from classifier}
}
\value{
patient_results, output
}
\description{
Takes individual spectra classification and converts to patient including sens/spec/npv/ppv
}
