% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC1_PC2_Plot.R
\name{pc_plots}
\alias{pc_plots}
\title{PC Plots}
\usage{
pc_plots(PCA_output, Labels)
}
\arguments{
\item{PCA_output}{Output from using PCA() function}
}
\value{
Data frame of PC1 and PC2 and their labels for plotting
}
\description{
Loadings and score plots ready to use
}
