% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Principle Component Analysis}
\usage{
PCA(spectra)
}
\arguments{
\item{spectra}{Processed spectral data}
}
\value{
PC loadings and scores in format PC#_score (e.g. pc1_score)

the full PCA object in the third list object
}
\description{
Computes PCA on spectra data
}
