% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polynomial_Baseline_Removal.R
\name{poly_baseline_removal}
\alias{poly_baseline_removal}
\title{Polynomial Baseline Removal}
\usage{
poly_baseline_removal(spectra, poly.order)
}
\arguments{
\item{spectra}{Interpolated spectra}

\item{poly.order}{Polynomial order}
}
\value{
spectra with polynomial baseline removed
}
\description{
Removes baseline contribution using hyperSpec package, more parameters to be added
}
