% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPV_NPV.R
\name{ppv_npv}
\alias{ppv_npv}
\title{NPV PPV Function}
\usage{
ppv_npv(conf_mat)
}
\arguments{
\item{conf_mat}{Confusion matrix output from random_forest_testing()}
}
\value{
PPV NPV of the model
}
\description{
To compute sens and spec on a confusion matrix output form RF
}
