% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_All_Spectra.R
\name{process_all_spec}
\alias{process_all_spec}
\title{Processing spectra en masse}
\usage{
process_all_spec()
}
\value{
A list, with first element the corrected and normalised

spectra, the other element contrains date + ID information
}
\description{
Runs all spectra in a file through the preprocessing sequence, outputting
a data frame of processed spectra.
}
