% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HISTORIC_PPSEQ.R
\name{process_all_spec_historic}
\alias{process_all_spec_historic}
\title{Old version of preprocess seq - smooths too, but not sure it shifts}
\usage{
process_all_spec_historic()
}
\description{
Old version of preprocess seq - smooths too, but not sure it shifts
}
