% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest100ave.R
\name{random_forest_ave}
\alias{random_forest_ave}
\title{Random Forest - Averaging 100}
\usage{
random_forest_ave(training_setin, testing_setin, cv_setin, test_set_ids,
  cv_set_ids)
}
\arguments{
\item{training_setin}{This is your training data, with the last column (V1016) to}

\item{testing_setin}{Testing set with V1016 containing classification independent test set}

\item{cv_setin}{Cross validation set}

\item{test_set_ids}{Vector containing testing set ids}

\item{cv_set_ids}{Vector with cross val ids}
}
\value{
prob_ave_cv, output_cv, prob_ave_test, output_test
}
\description{
Trains model using RF, also includes some QC removing data outside +/- 5 SD
}
