% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_Forest_Testing.R
\name{random_forest_testing}
\alias{random_forest_testing}
\title{Random Forest Testing}
\usage{
random_forest_testing(rf_model, rf_model_600, test_set, top600)
}
\arguments{
\item{rf_model}{Complete RF model}

\item{rf_model_600}{RF model using the top 600 features}

\item{test_set}{Data to test, last column should be the 1/0 cancer/control flag}

\item{top600}{Top 600 features as selected by RF model}
}
\value{
rf_table, rf_table_600 (confusions matrices), rf_perf, rf_perf_600 (ROC curves for each)
}
\description{
For use with random_forest_training to test models
}
