% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_Forest_Training.R
\name{random_forest_training}
\alias{random_forest_training}
\title{Random Forest Training}
\usage{
random_forest_training(training_setin, rmv_out)
}
\arguments{
\item{training_setin}{This is your training data, with the last column (V1016) to}

\item{rmv_out}{Choice to remove outliers > 5SD from data points (and if more than 5 points at >5SDs)}

\item{training_setin}{contain 1 or 0 for cancer/control respectively}
}
\value{
rf_model, rf_model_600 (uses top 600 features), top600 (top 600 features)
}
\description{
Trains model using RF, also includes some QC removing data outside +/- 5 SD
}
