% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCF.R
\name{RCF}
\alias{RCF}
\title{Rolling Circle Filter}
\usage{
RCF(raw_spec, samp_num, baseline_fit, radius)
}
\arguments{
\item{raw_spec}{Spectra data that has been interpolated already}

\item{samp_num}{Number of samples within data - user input}

\item{baseline_fit}{Choice of baseline fit from Pchip or Linear}

\item{radius}{RCF radius selection}
}
\value{
Corrected spectra with RCF applied (background removed)
}
\description{
Background removal for spectral data using RCF
}
