% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCF_Development.R, R/RCF_Efficiency_Testing.R
\name{RCF_dev}
\alias{RCF_dev}
\title{Rolling Circle Filter- Dev - use with apply}
\usage{
RCF_dev(raw_spec, baseline_fit, radius)

RCF_dev(raw_spec, baseline_fit, radius)
}
\arguments{
\item{raw_spec}{Spectra data that has been interpolated already}

\item{baseline_fit}{Choice of baseline fit from Pchip or Linear}

\item{radius}{RCF radius selection}

\item{raw_spec}{Spectra data that has been interpolated already}

\item{baseline_fit}{Choice of baseline fit from Pchip or Linear}

\item{radius}{RCF radius selection}
}
\value{
Corrected spectra with RCF applied (background removed)

Corrected spectra with RCF applied (background removed)
}
\description{
Background removal for spectral data using RCF

Background removal for spectral data using RCF
}
