% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCF_DVP_070819.R
\name{RCF_GENERALISED}
\alias{RCF_GENERALISED}
\title{Rolling Circle Filter- Dev - use with apply}
\usage{
RCF_GENERALISED(raw_spectra, radius, wavenumber)
}
\arguments{
\item{radius}{RCF radius selection}

\item{wavenumber}{Wavenumber range included}

\item{raw_spec}{Spectra data that has been interpolated already}
}
\value{
Corrected spectra with RCF applied (background removed)
}
\description{
Background removal for spectral data using RCF, Pchip used for baseline fit
}
