% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_in_shift1001_rc.R,
%   R/Read_in_shift1004_rc_REFLEX.R
\name{read_in_dev_axis}
\alias{read_in_dev_axis}
\title{Read In, Shift, Interp - Shifs to 1001 Post Neon Calibration RC applied}
\usage{
read_in_dev_axis(rc, sys = "R")

read_in_dev_axis(rc, sys = "R")
}
\value{
Interpolated and Shifted Spectra to 1001

Interpolated and Shifted Spectra to 1004
}
\description{
For use with data post the neon calibration in March 2019

For use with data before neon cal on reflex system data
}
