% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPC_Interp.R
\name{read_in_HPC_Shift}
\alias{read_in_HPC_Shift}
\title{HPC Shift Interp Function}
\usage{
read_in_HPC_Shift(file_dir)
}
\value{
Interpolated and Shifted Spectra to 1004
}
\description{
For use with HPC - no user interaction, points to file with data sets and processes
}
