% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_in_shift_interp_labelfree.R
\name{read_in_interp_labelfree}
\alias{read_in_interp_labelfree}
\title{Read In, Shift, Interp label free}
\usage{
read_in_interp_labelfree(file_dir)
}
\value{
Interpolated and Shifted Spectra without labels
}
\description{
Read In, Shift, Interp label free
}
