% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_in_no_shift.R
\name{read_in_noshift}
\alias{read_in_noshift}
\title{Read In, interp no shifting}
\usage{
read_in_noshift(file_dir)
}
\value{
Interpolated spectra without shifting
}
\description{
Read In, interp no shifting
}
