% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPC_Interp.R,
%   R/Read_In_Shift_Interpolate_Post_Neon_Cal.R
\name{read_in_post_neon_cal}
\alias{read_in_post_neon_cal}
\title{HPC Shift Interp Function}
\usage{
read_in_post_neon_cal()

read_in_post_neon_cal()
}
\value{
Interpolated and Shifted Spectra to 1004

Interpolated and Shifted Spectra to 1001
}
\description{
For use with HPC - no user interaction, points to file with data sets and processes

For use with data post the neon calibration in March 2019
}
