% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_In_Shift_Interpolate.R
\name{read_in_shift_interp}
\alias{read_in_shift_interp}
\title{Reading In, Shifting and Interpolating}
\usage{
read_in_shift_interp()
}
\description{
Reads in spectra, shifts to 1004 and interpolates on the x axis
}
