% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_in_shift_interp_manyspec.R
\name{read_in_shift_interp_manyspec}
\alias{read_in_shift_interp_manyspec}
\title{Reading In, Shifting and Interpolating for many spectra (e.g >100)}
\usage{
read_in_shift_interp_manyspec()
}
\description{
Reads in spectra, shifts to 1004 and interpolates on the x axis
Note currently phenyl peak to 2nd order poly rather than 4th due to error
}
