% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reading_in_and_Interpolating.R
\name{read_interp_spectra}
\alias{read_interp_spectra}
\title{Reading in raw spectra and interpolating}
\usage{
read_interp_spectra()
}
\value{
Interpolated spectra in hyperSpec format
}
\description{
Reads in mutiple raw spectra and interpolates using hyperSpec pkg
}
