% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repeat_Labels_Adhoc.R
\name{reps}
\alias{reps}
\title{Labels for repeats}
\usage{
reps(label)
}
\arguments{
\item{label}{a label for the data, eg 1,1,1,2,2,2}
}
\value{
Returns a data frame contating labels from repeats e.g 1.1, 1.2
}
\description{
For use in adhoc type analysis
}
