% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Savitsky_Golay_Filtering.R
\name{SavGol}
\alias{SavGol}
\title{Savitsky Golay filter}
\usage{
SavGol(spectra, poly_order)
}
\arguments{
\item{spectra}{Spectra data}

\item{poly_order}{Order of polynomial for filtering}
}
\value{
Filtered spectra
}
\description{
Smoothing using a savitsky golay, parameter is polynomial order
}
