% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity_Specificity.R
\name{sen_spec}
\alias{sen_spec}
\title{Sensitivity Specificity Function}
\usage{
sen_spec(conf_mat)
}
\arguments{
\item{conf_mat}{Confusion matrix output from random_forest_testing()}
}
\value{
sensitivity and specificity of the model
}
\description{
To compute sens and spec on a confusion matrix output form RF
}
