% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wavelet_Denoise.R
\name{wavelet_denoise}
\alias{wavelet_denoise}
\title{Wavelet Denoise
Smoothing using the wavelet denoise method}
\usage{
wavelet_denoise(spectra, filter_len)
}
\arguments{
\item{spectra}{Spectral object requiring denoising}

\item{filter_len}{Length of filter to be used with wavelet denoise. must be even and max of N=20}
}
\value{
Denoiseed spectra
}
\description{
Wavelet Denoise
Smoothing using the wavelet denoise method
}
\references{
Adapted from PRFFECT package
}
