% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Amide.R
\name{norm_a}
\alias{norm_a}
\title{Normalise at the largest amide peak}
\usage{
norm_a(spectra)
}
\arguments{
\item{spectra}{Spectra in matrix format for normalising}
}
\value{
Normalised spectra
}
\description{
Normalising function using min/max method at amide peak
}
