% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Phenylalanine.R
\name{norm_p}
\alias{norm_p}
\title{Normalise at Phenylalanine peak}
\usage{
norm_p(spectra, bin_size = 1)
}
\arguments{
\item{spectra}{Spectra in matrix format for normalising}

\item{bin_size}{Number of wavenumbers in a bin to shift region to normalise to accordingly}
}
\value{
Normalised spectra
}
\description{
Normalising function using min/max method at phenylalanine peak
}
