% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_Amide.R, R/Norm_Phenylalanine.R
\name{norm_p}
\alias{norm_p}
\title{Normalise at the largest amide peak}
\usage{
norm_p(spectra)

norm_p(spectra)
}
\arguments{
\item{spectra}{Spectra in matrix format for normalising}

\item{spectra}{Spectra in matrix format for normalising}
}
\value{
Normalised spectra

Normalised spectra
}
\description{
Normalising function using min/max method at amide peak

Normalising function using min/max method at phenylalanine peak
}
