% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opt_Dev_RCF.R, R/wrapper_func.R
\name{opt_process_hpc_dev}
\alias{opt_process_hpc_dev}
\title{Options for pre-processing spectra - DEV}
\usage{
opt_process_hpc_dev(spectra, norm_meth, rm_bl, baseline_fit, RCF_rad,
  poly_order)

opt_process_hpc_dev(spectra, norm_meth, rm_bl, baseline_fit, RCF_rad,
  poly_order)
}
\arguments{
\item{spectra}{Spectra data that has been interpolated already}

\item{norm_meth}{Choice of normalisation}

\item{RCF_rad}{RCF radius selection}

\item{spectra}{Spectra data that has been interpolated already}

\item{norm_meth}{Choice of normalisation}

\item{RCF_rad}{RCF radius selection}
}
\value{
Preprocessed spectra

Preprocessed spectra
}
\description{
Options for pre-processing spectra - DEV

Options for pre-processing spectra - developing version
}
