% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opt_Dev_RCF.R
\name{opt_process_hpc_dev}
\alias{opt_process_hpc_dev}
\title{Options for pre-processing spectra - DEV}
\usage{
opt_process_hpc_dev(spectra, norm_meth, RCF_rad)
}
\arguments{
\item{spectra}{Spectra data that has been interpolated already}

\item{norm_meth}{Choice of normalisation}

\item{RCF_rad}{RCF radius selection}
}
\value{
Preprocessed spectra
}
\description{
Options for pre-processing spectra - DEV
}
