% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_In_Shift_Interpolate_Post_Neon_Cal.R
\name{read_in_post_neon_cal}
\alias{read_in_post_neon_cal}
\title{Read In, Shift, Interp - Shifs to 1001 Post Neon Calibration}
\usage{
read_in_post_neon_cal()
}
\value{
Interpolated and Shifted Spectra to 1001
}
\description{
For use with data post the neon calibration in March 2019
}
