% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral_Plotting.R
\name{spectral_plots}
\alias{spectral_plots}
\title{Spectral Plotting}
\usage{
spectral_plots(raw_spec, samp_num, wavenumber)
}
\arguments{
\item{raw_spec}{Spectra in matrix form (like raw spectra)}

\item{samp_num}{# of samples}
}
\value{
plot of spectra
}
\description{
Easy and clear plotting for spectral data for processed and pre-processed spectra
}
